requirements.txt
langchain
openai
faiss-cpu
pandas
langchain-community
tiktoken
transformers
reportlab
sentence-transformers
flask